separate code for nick and quit commands
