+++ requirements/srcs/Nick.cpp
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   Commands.cpp                                       :+:      :+:    :+:   */
+/*   Nick.cpp                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: abellakr <abellakr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/03 22:24:11 by abellakr          #+#    #+#             */
-/*   Updated: 2023/06/04 01:58:09 by abellakr         ###   ########.fr       */
+/*   Updated: 2023/06/05 14:39:02 by abellakr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,26 +47,4 @@ void Server::nick(size_t pfdsindex)
             ERR_ERRONEUSNICKNAME(pfdsindex, MS[1]);
         }  
     }
-}
-
-
-bool Server::quit(size_t pfdsindex)
-{
-    if(MS.size() != 1)
-        return false; 
-    else
-    {
-        // remove client from map
-        std::cout << "client disconnected sockfd: " << pfds[pfdsindex].fd << std::endl; 
-        std::map<int,Client>::iterator itmap = ClientsMap.find(pfds[pfdsindex].fd);
-        ClientsMap.erase(itmap);
-        // remove client from pollfd vector
-        std::vector<pollfd>::iterator itvec = pfds.begin();
-        itvec += pfdsindex;
-        close(pfds[pfdsindex].fd);
-        pfds.erase(itvec);
-        // remove client from channel
-        return true;
-    }
-    return false;
 }
